Clear the rest of the reliability findings
WXPermissions.checkWXPermissions opened with a null guard on the player.
No caller can reach it with null -- all thirty-nine pass one from an
instanceof cast, an event, or a local derived from one -- so it never
fired. It was not harmless: Sonar reads it as proof the parameter is
nullable, and since every caller messages that same player in the else
branch, seven javabugs:S2259 reports across Build, Compass, Dial,
RingCommand, BeamCommand and GateInteractionHandler all traced back to
this one line. Removing it clears all seven, with no behaviour change
because the branch was unreachable.

No test: this deletes an unreachable branch, and the permission paths
themselves are already covered by OperatorPermissionTest.

removeStargate ignored the result of delete(), so a gate file that could
not be removed came back on the next load in silence; the config writer
ignored createNewFile the same way. Both report now.

Ten S2184 sites computed block coordinates in int and let the result
widen. None can overflow at Minecraft's coordinate limits, so the casts
are tidying rather than a fix -- they put the widening before the
arithmetic instead of after.

Left alone: five java:S3077 volatile-collection fields. Making those
thread-safe is a real concurrency decision about which of them are
touched off the main thread, not a mechanical change, and it wants its
own PR.

754 tests pass.

+++ src/main/java/com/wormhole_xtreme/wormhole/model/ring/Ring.java
@@ -698,8 +698,8 @@ public class Ring
      */
     public long anchorDistanceSquared(final Ring other)
     {
-        final long dx = anchorX - other.anchorX;
-        final long dz = anchorZ - other.anchorZ;
+        final long dx = (long) anchorX - other.anchorX;
+        final long dz = (long) anchorZ - other.anchorZ;
         return (dx * dx) + (dz * dz);
     }
